+++ package.json
@@ -19,6 +19,7 @@
     "babel-register": "~6.26.0",
     "chai": "~4.1.2",
     "expect": "~1.20.2",
+    "local-storage": "~1.4.2",
     "mocha": "~4.0.1"
   }
 }
