--- packages/kinvey-user/package.json
@@ -1,30 +1,0 @@
-{
-  "private": true,
-  "version": "3.9.3-beta.7",
-  "name": "kinvey-user",
-  "main": "src/index.js",
-  "dependencies": {
-    "es6-promise": "4.1.1",
-    "kinvey-acl": "^3.9.3-beta.7",
-    "kinvey-client": "^3.9.3-beta.7",
-    "kinvey-datastore": "^3.9.3-beta.7",
-    "kinvey-errors": "^3.9.3-beta.7",
-    "kinvey-identity": "^3.9.3-beta.7",
-    "kinvey-live": "^3.9.3-beta.7",
-    "kinvey-metadata": "^3.9.3-beta.7",
-    "kinvey-query": "^3.9.3-beta.7",
-    "kinvey-request": "^3.9.3-beta.7",
-    "kinvey-utils": "^3.9.3-beta.7",
-    "lodash": "4.17.4"
-  },
-  "devDependencies": {
-    "expect": "1.20.2",
-    "kinvey": "^3.9.3-beta.7",
-    "local-storage": "1.4.2",
-    "mocha": "4.0.1",
-    "nock": "9.0.28"
-  },
-  "scripts": {
-    "test": "mocha -s 250 test/**/*.spec.js"
-  }
-}
