--- packages/kinvey-user/src/index.js
@@ -1,1 +1,0 @@
-module.exports = require('./user');
